backend: Fix `post_login` handler
There were two things wrong. On the one hand the query string is in the
body of the http request as the method is post and we were expecting it
in the URL. On the other hand a successful login should result in a
redirect to `tracker` and we just responded with an `OK` status code.
This commit fixes both issues.