backend: Fix availability issue
A client not proceeding to a TLS negotiation after establishing a TCP
connection blocks the server. This occasionally happens due to the
interenet background noise. This commit sets a timeout for a TLS
negotiation preventing the DOS from background noise. However, the
timeout does not in general prevent DOS attacks as an attacker could
open a lot of TLS negotiations accummulating the timeouts.